Check if emotion chip skills can be used

+++ Source/relay/TourGuide/Items of the Month/2021/Emotion Chip.ash
@@ -12,60 +12,60 @@ void IOTMEmotionChipGenerateResource(ChecklistEntry [int] resource_entries)
 		string [int] emotions;
 				
         int emotionDisappointed = clampi(3 - get_property_int("_feelDisappointedUsed"), 0, 3);
-        if (emotionDisappointed > 0)
+        if (emotionDisappointed > 0 && $skill[Feel Disappointed].skill_is_usable())
 			{
             emotions.listAppend(emotionDisappointed + " Disappointments left. Your therapist has a lot of feelings about this one.");
 			}
         int emotionExcitement = clampi(3 - get_property_int("_feelExcitementUsed"), 0, 3);
-        if (emotionExcitement > 0)
+        if (emotionExcitement > 0 && $skill[Feel Excitement].skill_is_usable())
 			{
             emotions.listAppend(emotionExcitement + " Excitement left. 20 advs of +25 Mus/Mys/Mox.");
 			}
         int emotionLonely = clampi(3 - get_property_int("_feelLonelyUsed"), 0, 3);
-        if (emotionLonely > 0)
+        if (emotionLonely > 0 && $skill[Feel Lonely].skill_is_usable())
 			{
             emotions.listAppend(emotionLonely + " Lonelys left. 20 advs of -5% Combat.");
 			}
         int emotionLost = clampi(3 - get_property_int("_feelLostUsed"), 0, 3);
-        if (emotionLost > 0)
+        if (emotionLost > 0 && $skill[Feel Lost].skill_is_usable())
 			{
             emotions.listAppend(emotionLost + " Losts left. 20 advs of a weird Teleportitis buff.");
 			}
         int emotionNervous = clampi(3 - get_property_int("_feelNervousUsed"), 0, 3);
-        if (emotionNervous > 0)
+        if (emotionNervous > 0 && $skill[Feel Nervous].skill_is_usable())
 			{
             emotions.listAppend(emotionNervous + " Nervouses left. 20 advs of passive damage.");
 			}
         int emotionPeaceful = clampi(3 - get_property_int("_feelPeacefulUsed"), 0, 3);
-        if (emotionPeaceful > 0)
+        if (emotionPeaceful > 0 && $skill[Feel Peaceful].skill_is_usable())
 			{
             emotions.listAppend(emotionPeaceful + " Peacefuls left. 20 advs of +2 elemental resist.");
 			}
         int emotionPride = clampi(3 - get_property_int("_feelPrideUsed"), 0, 3);
-        if (emotionPride > 0)
+        if (emotionPride > 0 && $skill[Feel Pride].skill_is_usable())
 			{
             emotions.listAppend(emotionPride + " Prides left. Triple stat gain from current fight.");
 			}
         int emotionHatred = clampi(3 - get_property_int("_feelHatredUsed"), 0, 3);
-        if (emotionHatred > 0)
+        if (emotionHatred > 0 && $skill[Feel Hatred].skill_is_usable())
 			{
             emotions.listAppend(emotionHatred + " Hatreds left. 50-turn banish.");
 			
 			resource_entries.listAppend(ChecklistEntryMake("__skill feel hatred", "", ChecklistSubentryMake(pluralise(emotionHatred, "Feel Hatred", "Feels Hatreds"), "", "Cast Feel Hatred. Free run/banish.")).ChecklistEntrySetCombinationTag("banish").ChecklistEntrySetIDTag("Emotion chip feel hatred banish"));
 			}
         int emotionEnvy = clampi(3 - get_property_int("_feelEnvyUsed"), 0, 3);
-        if (emotionEnvy > 0)
+        if (emotionEnvy > 0 && $skill[Feel Envy].skill_is_usable())
 			{
             emotions.listAppend(emotionEnvy + " Envys left. Forces the monster to drop all items, but leaves killing it up to you.");
 			}
         int emotionNostalgic = clampi(3 - get_property_int("_feelNostalgicUsed"), 0, 3);
 		monster nostalgicMonster = (get_property_monster("lastCopyableMonster"));
-        if (emotionNostalgic > 0)
+        if (emotionNostalgic > 0 && $skill[Feel Nostalgic].skill_is_usable())
 			{
             emotions.listAppend(emotionNostalgic + " Nostalgias left. Copy a drop table. Can currently feel nostalgic for: " + HTMLGenerateSpanFont(nostalgicMonster, "blue"));
 			}
         int emotionSuperior = clampi(3 - get_property_int("_feelSuperiorUsed"), 0, 3);
-        if (emotionSuperior > 0)
+        if (emotionSuperior > 0 && $skill[Feel Superior].skill_is_usable())
 			{
             emotions.listAppend(emotionSuperior + " Superiors left. +1 PvP Fight if used as killshot.");
 			}
